Restart DMA transfers on calls to log
If there is a completed transfer, calls to log will complete the
transfer, enqueue the newly-added data, and re-schedule the transfer.
This reduces the number of poll() calls required to send data to a
host.

Previously, log usages like

  log::info!("Hey");
  log::info!("there");

would immediately schedule "Hey" to be written. Then, "there" would
be scheduled to be written on the next call to poll(). As of this
commit, if the transfer of "Hey" completes before writing "there,"
the implementation will immediately schedule "there" to be written.